+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.3.28 (05-12-2024):
+
+Try extract TraceContext from "traceparent" header.
+
+## 0.3.27 (21-11-2024):
+
+Fix TracingMiddleware not adding trace id header after ExceptionHandlerMiddleware
+
+## 0.3.26 (21-11-2024):
+
+Update ClusterClient library.
+
 ## 0.3.25 (29-10-2024):
 
 Rebuild nuget package to use new version of Hercules.Client.Abstractions